Extract shared pagination from Triples
